MenuShell: activate a menu item before deactivating its parent
If we have a standalone menu that is destroyed on deactivation and that should
perform an action if activated before, that won't work.

The reason is that gtk_menu_shell_activate_item deactivates the menu item
container before activating the activated item and don't see any rationale for
what is currently done.

https://bugzilla.gnome.org/show_bug.cgi?id=689247