a11y: Redo vfuncs of GtkContainerAccessible
Rename from add_gtk and remove_gtk to simply add and remove and make
them take accessible and widget as parameters.
We do pass in the child widget and not its accessible to give the vfunc
the chance to actually figure out if it needs to create the accessible
for a given widget or not.