range: don't trim the trough rectange by the trough margin
Commit e32da246a84492f02b4dac5a448bb46cb23d68b8 made GtkRange's trough
respect the CSS margin property, but it also trimmed the box in which
the trough reacts to click events by the margin.
We still want to catch events in that area instead, and just make sure
the margin is applied when drawing (which was already implemented by
that commit).

This commit reverts the parts of
e32da246a84492f02b4dac5a448bb46cb23d68b8 that didn't involve drawing,
fixing the bug.

https://bugzilla.gnome.org/show_bug.cgi?id=691677