Fix gtk3-demo Visual Studio projects
Since we are linking in the resource items by the source, we need to
disable WholeProgramOptimization so that the resource stuff does get linked
into the demo binaries, so that they can be loaded properly.

Also make sure that gtk3-demo-application is also built with the multibyte
character set, like the rest.

This should fix bug 694342, at least for Visual Studio builds.